requirements.txt
pandas
numpy
scikit-learn
yfinance
python-binance
python-dotenv